Small refactor: separate app and components
